storage/sql: Rename sqlite storage package to more generic sql
