+++ pkg/lock-manager/config.go
@@ -3,13 +3,13 @@ package lockmanager
 import (
 	"github.com/heathcliff26/fleetlock/pkg/lock-manager/errors"
 	"github.com/heathcliff26/fleetlock/pkg/lock-manager/storage/redis"
-	"github.com/heathcliff26/fleetlock/pkg/lock-manager/storage/sqlite"
+	"github.com/heathcliff26/fleetlock/pkg/lock-manager/storage/sql"
 )
 
 type StorageConfig struct {
-	Type   string               `yaml:"type"`
-	SQLite *sqlite.SQLiteConfig `yaml:"sqlite,omitempty"`
-	Redis  *redis.RedisConfig   `yaml:"redis,omitempty"`
+	Type   string             `yaml:"type"`
+	SQLite *sql.SQLiteConfig  `yaml:"sqlite,omitempty"`
+	Redis  *redis.RedisConfig `yaml:"redis,omitempty"`
 }
 
 type Groups map[string]GroupConfig
